IBiSco: Job info

+++ scripts/new_ibisco_simulation.sh
@@ -45,6 +45,7 @@ cd Dati_$3/
 count=0
 for ((i=1; i<=$1; i++)); do
     num_tasks="$2"
+    while :; do
 
         srun --job-name="${job_name}_J${i}" -p parallel -n $num_tasks a.out > srun.log 2>&1 &
         sleep 1
@@ -85,30 +86,34 @@ for ((i=1; i<=$1; i++)); do
                 ;;
             "PD")
                 # Verifica se la reason è Resources, Priority o altro
-                while (( $(squeue -j $job_id -o "%R" -h) == *"Resources"* )); do
-                    echo "Il job ${job_name}_J${i} ha priorità massima ma non ha le risorse necessarie."
-                    scancel $job_id
+                if [[ "$(squeue -j $job_id -o "%R" -h)" == *"Resources"* ]]; then
+                    # Verifica se il job_id è il primo nella lista di priorità
+                    if [ "$job_id" == "$(sprio -S '-Y' | awk 'NR==2 {print $1}')" ]; then
 
-                    echo "Riduzione del numero di task di 5 e rilancio del job ${job_name}_J${i}..."
-                    ((num_tasks -= 5))
-                    srun --job-name="${job_name}_J${i}" -p parallel -n $num_tasks a.out > srun.log 2>&1 &
-                    sleep 5
-                    job_id=$(squeue -u $USER -n "${job_name}_J${i}" -o "%i" -h | head -n 1)
+                        echo "Il job ${job_name}_J${i} ha priorità massima ma non ha le risorse necessarie."
+                        scancel $job_id
+                        echo "Riduzione del numero di task di 5 e rilancio del job ${job_name}_J${i}..."
+                        ((num_tasks -= 5))
 
-                    if (( $num_tasks < 1 )); then
-                        echo "Raggiunto il minimo numero di task (-n 2) per job. Cancellazione del job ${job_name}_J${i}..."
+                        if (( $num_tasks < 1 )); then
+                            echo "Raggiunto il minimo numero di task (-n 2) per job. Cancellazione del job ${job_name}_J${i}..."
+                            scancel $job_id
+                            ((i--))
+                            ((count++))
+                            # esito+=("Cancellato (assenza di risorse)")
+                            # tasks_per_job+=(0)
+                            break
+                        fi
+                    else
+                        echo "Il Job ${job_name}_J${i} ha una priority troppo bassa. Cancellazione del job ${job_name}_J${i}..."
                         scancel $job_id
                         ((i--))
                         ((count++))
-                        # esito+=("Cancellato (assenza di risorse)")
-                        # tasks_per_job+=(0)
                         break
                     fi
-                done
 
                 # Controlla se il job è in priority
-                # elif [[ "$(squeue -j $job_id -o "%R" -h)" == *"Priority"* || "$(squeue -j $job_id -o "%R" -h)" == *"DOWN, DRAINED"* ]]; then
-                while (( $(squeue -j $job_id -o "%R" -h) != *"ibiscohpc"* )); do
+                elif [[ "$(squeue -j $job_id -o "%R" -h)" == *"Priority"* || "$(squeue -j $job_id -o "%R" -h)" == *"DOWN, DRAINED"* ]]; then
                     job_reason=$(squeue -j $job_id -o "%R" -h)
                     echo "Il job ${job_name}_J${i} con ID $job_id è in attesa (PD) con reason: $job_reason."
                     time=1800
@@ -143,9 +148,10 @@ for ((i=1; i<=$1; i++)); do
                         # tasks_per_job+=(0)
                         break
                     fi
-                done
+                fi
                 ;;
         esac
+    done
     jobs+=("${job_name}_J${i}")
     ids+=("${job_id}")
 done
